requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
face-recognition>=1.3.0
opencv-python>=4.8.1.78
numpy>=1.24.0
supabase>=2.0.3
python-dotenv>=1.0.0
pyjwt>=2.8.0
requests>=2.31.0
